chore(config): renamed config folder
