+++ src/config/properties.ts
