+++ src/config/properties/base.ts
