+++ src/config/web/server.ts
@@ -1,6 +1,6 @@
 import Fastify from "fastify";
-import { properties } from "@/configurations/properties";
-import { prisma } from "@/configurations/orm/prisma";
+import { properties } from "@/config/properties";
+import { prisma } from "@/config/orm/prisma";
 
 export const app = Fastify({ logger: true });
